requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
python-dotenv==1.0.0

